rest of ws tests reinstated
